refactor: fix capitalization in SimpleFactor.ts -> simpleFactor.ts
  Since it is not exporting a type or class it should starte with a lower
  case letter.
