+++ src/sequences/simpleFactor.ts
